feat(abclass): create facade index.js with full delegation and wire setupGlobals.js
Sections 9-10 of ACTION_PLAN.md — facade creation + test harness wiring.

Created src/backend/y_controllers/ABClassController/index.js facade:
- Constructor wires 5 sub-classes per spec Decision 4
- 8 public methods with verbatim bodies
- 22 private method one-line delegators with explicit parameter names
- persistAssignmentRun orchestrated as facade method
- module.exports = ABClassController preserves all existing require paths

Module-resolution conflict (Option A):
- Renamed ABClassController.js to ABClassController.legacy.js
- Created re-export shim at original path
- Suppressed pre-existing lint in legacy file (temporary)

Updated tests/setupGlobals.js with 5 ABClass* sub-class globals.

All 159 regression tests pass. Lint: 0 errors, 17 pre-existing warnings.
